refactor: move `mock_request` to `classes/mock`
