+++ tests/classes/__init__.py
@@ -1,3 +1,4 @@
 from __future__ import annotations
 
+from .mock import mock_request
 from .mock import MockResponse
